feat: update eslint and jest configuration
