+++ .eslintrc.json
@@ -1,0 +1,42 @@
+{
+  "root": true,
+  "env": {
+    "jest": true,
+    "node": true,
+    "es6": true
+  },
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "project": "./tsconfig_test.json"
+  },
+  "extends": [
+    "eslint:recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:@typescript-eslint/recommended-requiring-type-checking",
+    "prettier"
+  ],
+  "plugins": ["@typescript-eslint", "import", "unused-imports"],
+  "rules": {
+    "prefer-const": "error",
+    "no-var": "error",
+    "no-console": "off",
+    "sort-imports": "off",
+    "import/order": [
+      "error",
+      {
+        "alphabetize": {
+          "order": "asc",
+          "caseInsensitive": true
+        }
+      }
+    ],
+    "class-methods-use-this": "error",
+    "unused-imports/no-unused-imports": "error",
+    "@typescript-eslint/ban-ts-comment": "off",
+    "@typescript-eslint/no-unused-vars": "off"
+  },
+  "ignorePatterns": [
+    "lib/**/*",
+    "node_modules"
+  ]
+}
