+++ package.json
@@ -25,7 +25,7 @@
     "@graphql-tools/merge": "^8.4.0",
     "amplify-appsync-simulator": "^2.4.1",
     "amplify-nodejs-function-runtime-provider": "^2.3.13",
-    "aws-sdk": "^2.1340.0",
+    "aws-sdk": "^2.1341.0",
     "axios": "^1.3.4",
     "cfn-resolver-lib": "^1.1.8",
     "dataloader": "^2.2.2",
@@ -53,13 +53,14 @@
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-unused-imports": "^2.0.0",
     "jest": "^29.5.0",
     "prettier": "^2.8.6",
-    "semantic-release": "20.1.3",
+    "semantic-release": "^20.1.3",
     "serverless": "^3.28.1",
     "ts-jest": "^29.0.5",
     "ts-node": "^10.9.1",
-    "typescript": "^5.0.2"
+    "typescript": "^4.9.5"
   },
   "peerDependencies": {
     "serverless": "3"
