+++ src/__tests__/config.ts
@@ -1,0 +1,8 @@
+import Serverless from 'serverless/lib/Serverless';
+
+describe('resolveConfiguration', () => {
+  it('should generate a valid config', () => {
+    const serverless: Serverless = {
+    };
+  });
+});
