+++ src/__tests__/data-loaders/ElasticDataLoader.ts
@@ -1,31 +1,26 @@
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import { PassThrough } from 'stream';
 import * as AWS from 'aws-sdk';
 import axios from 'axios';
 import ElasticDataLoader from '../../data-loaders/ElasticDataLoader';
 
 describe('data-loaders/ElasticDataLoader', () => {
-  beforeEach(() => {
-    jest.spyOn(AWS.HttpClient.prototype, 'handleRequest');
-    jest.spyOn(axios, 'request');
-  });
-
-  afterEach(() => {
-    AWS.HttpClient.prototype.handleRequest.mockClear();
-    axios.request.mockClear();
-  });
-
   it('should send a request', async () => {
+    const spyRequest = jest.spyOn(axios, 'request');
+    spyRequest.mockImplementation(async () => {
+      return Promise.resolve({ data: { hits: {} } });
+    });
+
     const loader = new ElasticDataLoader({
       endpoint: 'https://my-elasticsearch-cluster.region.amazonaws.com',
-    });
-    axios.request.mockImplementation(async () => {
-      return { data: { hits: {} } };
     });
     const req = {
       path: '[index]/_search',
       operation: 'GET',
       params: {
         headers: {},
+        queryString: '',
         body: '{"query": { "match_all": {} }}',
       },
     };
@@ -34,6 +29,21 @@ describe('data-loaders/ElasticDataLoader', () => {
   });
 
   it('should send a signed request', async () => {
+    const spyHandleRequest = jest.spyOn(
+      // @ts-ignore
+      AWS.HttpClient.prototype,
+      'handleRequest',
+    );
+
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    let signedRequest: any;
+    const mockStream = new PassThrough();
+    spyHandleRequest.mockImplementation((request, _options, callback) => {
+      signedRequest = request;
+      // @ts-ignore
+      callback(mockStream);
+    });
+
     const loader = new ElasticDataLoader({
       endpoint: 'https://my-elasticsearch-cluster.region.amazonaws.com',
       useSignature: true,
@@ -41,20 +51,13 @@ describe('data-loaders/ElasticDataLoader', () => {
       secretAccessKey: 'fakeSecretAccessKey',
       region: '',
     });
-    const mockStream = new PassThrough();
-    let signedRequest;
-    AWS.HttpClient.prototype.handleRequest.mockImplementation(
-      (request, _options, callback) => {
-        signedRequest = request;
-        callback(mockStream);
-      },
-    );
     const body = '{"query": { "match_all": {} }}';
     const req = {
       path: '[index]/_search',
       operation: 'GET',
       params: {
         headers: {},
+        queryString: '',
         body,
       },
     };
