+++ src/data-loaders/ElasticDataLoader.ts
@@ -1,5 +1,11 @@
-import axios from 'axios';
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/no-unsafe-argument */
+/* eslint-disable @typescript-eslint/no-unsafe-return */
+/* eslint-disable @typescript-eslint/no-unsafe-call */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import * as AWS from 'aws-sdk';
+import axios from 'axios';
 
 export default class ElasticDataLoader {
   private config: any;
