+++ src/data-loaders/HttpDataLoader.ts
@@ -1,3 +1,7 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+/* eslint-disable @typescript-eslint/restrict-template-expressions */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
 import axios, { CustomParamsSerializer } from 'axios';
 import { isObject, forEach } from 'lodash';
