+++ src/getAppSyncConfig.ts
@@ -1,3 +1,5 @@
+import { forEach, merge } from 'lodash';
+import { AppSyncConfigInput } from './types/appSync';
 import {
   ApiKeyConfig,
   AppSyncConfig,
@@ -5,66 +7,26 @@ import {
   PipelineFunctionConfig,
   ResolverConfig,
 } from './types/plugin';
-import { A, O } from 'ts-toolbelt';
-import { forEach, merge } from 'lodash';
 
-/* Completely replaces keys of O1 with those of O */
-type Replace<O extends object, O1 extends object> = O.Merge<
-  O,
-  O.Omit<O1, A.Keys<O>>
->;
-
-export type DataSourceConfigInput = O.Omit<DataSourceConfig, 'name'>;
-
-export type ResolverConfigInput = O.Update<
-  O.Update<
-    O.Optional<ResolverConfig, 'type' | 'field'>,
-    'dataSource',
-    string | DataSourceConfigInput
-  >,
-  'functions',
-  (string | FunctionConfigInput)[]
->;
-
-export type FunctionConfigInput = Replace<
-  { dataSource: string | DataSourceConfigInput },
-  O.Omit<PipelineFunctionConfig, 'name'>
->;
-
-export type AppSyncConfigInput = Replace<
-  {
-    schema?: string | string[];
-    apiKeys?: (ApiKeyConfig | string)[];
-    resolvers?:
-      | Record<string, ResolverConfigInput>[]
-      | Record<string, ResolverConfigInput>;
-    pipelineFunctions?:
-      | Record<string, FunctionConfigInput>[]
-      | Record<string, FunctionConfigInput>;
-    dataSources:
-      | Record<string, DataSourceConfigInput>[]
-      | Record<string, DataSourceConfigInput>;
-  },
-  O.Optional<AppSyncConfig, 'additionalAuthentications'>
->;
 
 const flattenMaps = <T>(
   input?: Record<string, T> | Record<string, T>[],
 ): Record<string, T> => {
   if (Array.isArray(input)) {
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-return
     return merge({}, ...input);
   } else {
     return merge({}, input);
   }
 };
 
-export const isUnitResolver = (resolver: {
+const isUnitResolver = (resolver: {
   kind?: 'UNIT' | 'PIPELINE';
 }): resolver is { kind: 'UNIT' } => {
   return resolver.kind === 'UNIT';
 };
 
-export const isPipelineResolver = (resolver: {
+const isPipelineResolver = (resolver: {
   kind?: 'UNIT' | 'PIPELINE';
 }): resolver is { kind: 'PIPELINE' } => {
   return !resolver.kind || resolver.kind === 'PIPELINE';
