+++ tsconfig_test.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["./**/*.ts"],
+  "exclude": ["node_modules"]
+}
